minor fixes to housing counselor pdf view
This commit syncs the location of the housing counselor PDF view
template with its standard HTML template, both now located under
`cfgov/legacy/templates/hud`. I made some slight corrections to the PDF
template as well so that it properly includes the city and zipcode of
the counselors being listed.

This change also moves PDF HTML rendering solely within this
cfgov-refresh project, so that the external django-hud project is now
only relied on for JSON results. See also this PR:

https://github.com/cfpb/django-hud/pull/18

Relevant URLs for this functionality are now:

Base search: /find-a-housing-counselor/
HTML results: /find-a-housing-counselor/?zipcode=18020
HTML results for PDF: /find-a-housing-counselor/?zipcode=18020&pdf
PDF download: /save-hud-counselors-list/?zip=18020
